requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.128.8
frozenlist==1.8.0
h11==0.16.0
idna==3.11
multidict==6.7.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
SQLAlchemy==2.0.48
starlette==0.49.3
twilio==9.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
yarl==1.22.0
